feat: move user message to prmopt class
