+++ packages/vmind/src/base/tools/prompt/index.ts
@@ -3,7 +3,7 @@ import { IPrompt } from './types';
  * Prompt represents a LLM Prompt
  * Pass in Template when initialization
  * getPrompt method generates specific Prompt according to Template and Context
- * The subclass needs to rewrite the getprompt method to generate a specific Prompt
+ * The subclass needs to rewrite the getSystemPrompt and getUserPrompt method to generate a specific Prompt
  */
 export class Prompt<Context> implements IPrompt<Context> {
   template: string;
@@ -11,7 +11,10 @@ export class Prompt<Context> implements IPrompt<Context> {
   constructor(template: string) {
     this.template = template;
   }
-  getPrompt(context: Context) {
+  getSystemPrompt(context: Context) {
+    return this.template;
+  }
+  getUserPrompt(context: Context) {
     return this.template;
   }
 }
